added logs and changed computing of progress percentage in training progress bar
