+++ CODIGO/app/activity_recognition/app/src/main/java/org/activityrecognition/external/client/model/ModelClientFactory.java
@@ -4,8 +4,8 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class ModelClientFactory {
     private static ModelClient client;
-    //private static final String BASE_URL = "http://18.221.185.88:8080";
-    private static final String BASE_URL = "http://192.168.0.8:8080";
+    private static final String BASE_URL = "http://18.221.185.88:8080";
+    //private static final String BASE_URL = "http://192.168.0.8:8080";
 
     public static ModelClient getClient() {
         if (client == null) {
