+++ CODIGO/app/activity_recognition/app/src/main/res/layout/activity_main.xml
@@ -84,13 +84,6 @@
                 android:padding="12dp"
                 />
 
-            <TextView
-                android:id="@+id/txt_train_status"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="left"
-                android:text="" />
-
             <Button
                 android:id="@+id/btn_predict"
                 android:layout_width="fill_parent"
